requirements.txt
asgiref==3.7.2
bleach==6.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
Django==4.2.1
djangorestframework==3.14.0
docutils==0.20.1
fonttools==4.39.4
idna==3.4
importlib-metadata==6.6.0
jaraco.classes==3.2.3
joblib==1.2.0
keyring==23.13.1
kiwisolver==1.4.4
markdown-it-py==2.2.0
matplot==0.1.9
matplotlib==3.7.1
mdurl==0.1.2
more-itertools==9.1.0
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pkginfo==1.9.6
Pygments==2.15.1
pyloco==0.0.139
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pywin32-ctypes==0.2.0
readme-renderer==37.3
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.3.5
scikit-learn==1.2.2
scipy==1.10.1
SimpleWebSocketServer==0.1.2
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.1.0
twine==4.0.2
typing==3.7.4.3
tzdata==2023.3
urllib3==2.0.2
ushlex==0.99.1
webencodings==0.5.1
websocket-client==1.5.2
zipp==3.15.0
